+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.0.4 (08/28/21)
+
+- Introduced the `WikiLens: Ignored Files` setting
+- Updated the backlink node tooltip to display the page's name and line preview
+- Fixed a bug where the wikilink decorator was incorrectly styling text
+- Fixed a bug where the backlinks view within files was displaying a confusing `Start discussion` label
+
 ## v0.0.3 (08/21/21)
 
 - Fixed a bug in github.dev, that would duplicate files in the `Wiki` tree whenever you would edit them
@@ -12,3 +19,5 @@
 ## v0.0.1 (08/20/21)
 
 Initial release! 🚀
+
+[[🔎 WikiLens]]